[ENH] Collection transformer `compose.py` cleanup (#683)
* tidy

* notebook

* docs

* docs
