+++ aeon/transformations/compose.py
@@ -4,6 +4,7 @@
 
 from warnings import warn
 
+import numpy as np
 import pandas as pd
 from sklearn import clone
 from sklearn.utils.metaestimators import if_delegate_has_method
@@ -23,6 +24,7 @@ from aeon.utils.validation.series import check_series
 __author__ = ["fkiraly", "mloning", "miraep8", "aiwalter", "SveaMeyer13"]
 __all__ = [
     "ColumnwiseTransformer",
+    "ColumnConcatenator",
     "FeatureUnion",
     "FitInTransform",
     "Id",
@@ -1635,6 +1637,58 @@ def _check_is_pdseries(z):
         z = z.to_frame()
         is_series = True
     return z, is_series
+
+
+class ColumnConcatenator(BaseTransformer):
+    """Concatenate multivariate series to a long univariate series.
+
+    Transformer that concatenates multivariate time series/panel data
+    into long univariate time series/panel
+        data by simply concatenating times series in time.
+    """
+
+    _tags = {
+        "scitype:transform-input": "Series",
+        # what is the scitype of X: Series, or Panel
+        "scitype:transform-output": "Series",
+        # what scitype is returned: Primitives, Series, Panel
+        "scitype:instancewise": False,  # is this an instance-wise transform?
+        "X_inner_mtype": ["pd-multiindex", "pd_multiindex_hier"],
+        # which mtypes do _fit/_predict support for X?
+        "y_inner_mtype": "None",  # which mtypes do _fit/_predict support for X?
+        "fit_is_empty": True,  # is fit empty and can be skipped? Yes = True
+    }
+
+    def _transform(self, X, y=None):
+        """Transform the data.
+
+        Concatenate multivariate time series/panel data into long
+        univariate time series/panel
+        data by simply concatenating times series in time.
+
+        Parameters
+        ----------
+        X : nested pandas DataFrame of shape [n_samples, n_features]
+            Nested dataframe with time-series in cells.
+
+        Returns
+        -------
+        Xt : pandas DataFrame
+          Transformed pandas DataFrame with same number of rows and single
+          column
+        """
+        Xst = pd.DataFrame(X.stack())
+        Xt = Xst.swaplevel(-2, -1).sort_index().droplevel(-2)
+
+        # the above has the right structure, but the wrong indes
+        # the time index is in general non-unique now, we replace it by integer index
+        inst_idx = Xt.index.get_level_values(0)
+        t_idx = [range(len(Xt.loc[x])) for x in inst_idx.unique()]
+        t_idx = np.concatenate(t_idx)
+
+        Xt.index = pd.MultiIndex.from_arrays([inst_idx, t_idx])
+        Xt.index.names = X.index.names
+        return Xt
 
 
 class YtoX(BaseTransformer):
